Use uint32_t for compilation unit index instead of uint16_t.

+++ src/compiler_llvm/utils_llvm.h
@@ -32,7 +32,7 @@ namespace art {
 #define VERIFY_LLVM_FUNCTION(func)
 #endif
 
-inline static std::string ElfFuncName(uint16_t idx) {
+inline static std::string ElfFuncName(uint32_t idx) {
   return StringPrintf("Art%u", static_cast<unsigned int>(idx));
 }
